revno: 4801 message:   Validate for patientChart when creating/updating a chart; Check duplicate name of other object is wrong (Patient module).
committer: Tran Chau <tran.hispvietnam@gmail.com>
branch nick: dhis2
timestamp: Mon 2011-10-03 16:40:38 +0700

+++ dhis-2/dhis-web/dhis-web-maintenance/dhis-web-maintenance-patient/src/main/java/org/hisp/dhis/patient/action/dataentryform/ValidateDataEntryFormAction.java
@@ -96,7 +96,7 @@ public class ValidateDataEntryFormAction
     {
         DataEntryForm match = dataEntryFormService.getDataEntryFormByName( name );
 
-        if ( match != null && (dataEntryFormId == null || match.getId() != dataEntryFormId) )
+        if ( match != null && (dataEntryFormId == null || match.getId() != dataEntryFormId.intValue()) )
         {
             message = i18n.getString( "duplicate_names" );
